requirements/requirements.txt
awscli==1.12.0
awsebcli==3.12.1
blessed==1.14.2
botocore==1.8.0
cement==2.8.2
colorama==0.3.7
Django==1.11.7
django-extensions==1.9.9
django-storages==1.6.5
djangorestframework==3.7.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
jmespath==0.9.3
pathspec==0.5.0
Pillow==5.0.0
psycopg2==2.7.3.2
pyasn1==0.3.7
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.9.1
rsa==3.4.2
s3transfer==0.1.11
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
typing==3.6.2
wcwidth==0.1.7
websocket-client==0.44.0
urllib3==1.22
django-s3-storage==0.12.3
boto==2.48.0
djangorestframework-jwt==1.11.0
google-auth==1.4.1
google-api-python-client==1.6.5
google-auth-oauthlib
django-cors-headers==2.1.0
cryptography